requirements.txt
Flask>=2.2.0
APScheduler>=3.10.0
requests>=2.28.1
beautifulsoup4>=4.11.1
pytz>=2023.3
python-dateutil>=2.8.2
python-dotenv